+++ frontends/mit-open/src/components/RestrictedRoute/RestrictedRoute.test.tsx
