+++ frontends/mit-open/src/components/RestrictedRoute/RestrictedRoute.tsx
