+++ frontends/mit-open/src/pages/ErrorPage/ErrorPage.test.tsx
@@ -6,7 +6,7 @@ import type { TestAppOptions } from "../../test-utils"
 import ErrorPage from "./ErrorPage"
 import { setMockResponse, mockAxiosInstance as axios } from "api/test-utils"
 import { allowConsoleErrors } from "ol-util/test-utils"
-import RestrictedRoute from "../../components/RestrictedRoute"
+import RestrictedRoute from "components/RestrictedRoute/RestrictedRoute"
 import { Permissions } from "../../util/permissions"
 
 /**
