+++ frontends/mit-open/src/pages/InfinitePages/pages/EditFieldPage.tsx
@@ -5,7 +5,7 @@ import { Container, Tab, TabList, TabContext, TabPanel } from "ol-design"
 
 import { MetaTags } from "ol-util"
 
-import { GridColumn, GridContainer } from "../../../components/layout"
+import { GridColumn, GridContainer } from "components/GridLayout/GridLayout"
 import { useFieldDetails } from "../../../api/fields"
 import EditFieldAppearanceForm from "../page-components/EditFieldAppearanceForm"
 import EditFieldBasicForm from "../page-components/EditFieldBasicForm"
